# server/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/tournaments": {
      "get": {
        "operationId": "TournamentController_findAll",
        "parameters": [],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TournamentDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Tournament"
        ]
      },
      "post": {
        "operationId": "TournamentController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTournamentDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TournamentDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Tournament"
        ]
      }
    },
    "/api/tournaments/{id}": {
      "put": {
        "operationId": "TournamentController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTournamentDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TournamentDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Tournament"
        ]
      },
      "delete": {
        "operationId": "TournamentController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Tournament"
        ]
      }
    },
    "/api/tournaments/{tournamentId}/boxers": {
      "get": {
        "operationId": "TournamentController_getBoxersForTournament",
        "parameters": [
          {
            "name": "tournamentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BoxerGetDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Tournament"
        ]
      }
    },
    "/api/tournaments/{tournamentId}/opponents/{boxerId}": {
      "get": {
        "operationId": "TournamentController_getPossibleOpponents",
        "parameters": [
          {
            "name": "tournamentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "boxerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OpponentGetDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Tournament"
        ]
      }
    },
    "/api/tournaments/{tournamentId}/fights": {
      "get": {
        "operationId": "TournamentController_getFightsByTournamentId",
        "parameters": [
          {
            "name": "tournamentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FightGetDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Tournament"
        ]
      }
    },
...
  },
  "info": {
    "title": "Battlecard",
    "description": "The Fightmaker App",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "TournamentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique tournament ID"
          },
          "name": {
            "type": "string",
            "description": "Tournament name"
          },
          "userId": {
            "type": "string",
            "description": "User who created the tournament"
          },
          "date": {
            "type": "string",
            "description": "Tournament date (YYYY-MM-DD)"
          },
          "address": {
            "type": "string",
            "description": "Tournament address (optional)"
          },
          "zipCode": {
            "type": "string",
            "description": "Tournament zip code (optional)"
          },
          "city": {
            "type": "string",
            "description": "Tournament city (optional)"
          },
          "formattedAddress": {
            "type": "string",
            "description": "Formatted address combining street, city, and zipCode"
          }
        },
        "required": [
          "id",
          "name",
          "userId",
          "date"
        ]
      },
      "CreateTournamentDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Tournament name"
          },
          "date": {
            "type": "string",
            "description": "Tournament date (YYYY-MM-DD)"
          },
          "address": {
            "type": "string",
            "description": "Tournament address (optional)"
          },
          "zipCode": {
            "type": "string",
            "description": "Tournament zip code (optional)"
          },
          "city": {
            "type": "string",
            "description": "Tournament city (optional)"
          }
        },
        "required": [
          "name",
          "date"
        ]
      },
      "UpdateTournamentDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Tournament name"
          },
          "date": {
            "type": "string",
            "description": "Tournament date (YYYY-MM-DD)"
          },
          "address": {
            "type": "string",
            "description": "Tournament address (optional)"
          },
          "zipCode": {
            "type": "string",
            "description": "Tournament zip code (optional)"
          },
          "city": {
            "type": "string",
            "description": "Tournament city (optional)"
          }
        },
        "required": [
          "name",
          "date"
        ]
      },
      "Gender": {
        "type": "string",
        "enum": [
          "female",
          "male"
        ],
        "description": "Gender"
      },
      "BoxerGetDto": {
        "type": "object",
        "properties": {
          "gender": {
            "description": "Gender",
            "allOf": [
              {
                "$ref": "#/components/schemas/Gender"
              }
            ]
          },
          "id": {
            "type": "string",
            "description": "Unique boxer ID"
          },
          "lastName": {
            "type": "string",
            "description": "Boxer's last name"
          },
          "firstName": {
            "type": "string",
            "description": "Boxer's first name"
          },
          "birthDate": {
            "type": "string",
            "description": "Date of birth (YYYY-MM-DD)"
          },
          "nbFights": {
            "type": "number",
            "description": "Number of past fights of the boxer"
          },
          "club": {
            "type": "string",
            "description": "Club name"
          },
          "weight": {
            "type": "number",
            "description": "Weight in kg (optional)"
          },
          "license": {
            "type": "string",
            "description": "License number"
          },
          "userId": {
            "type": "string",
            "description": "User who created this boxer"
          },
          "created": {
            "type": "string",
            "description": "Creation timestamp"
          },
          "updated": {
            "type": "string",
            "description": "Update timestamp"
          },
          "category": {
            "type": "string",
            "description": "Weight/age category"
          },
          "categoryShort": {
            "type": "string",
            "description": "Short category label"
          },
          "eligibleFights": {
            "type": "number",
            "description": "Number of eligible fights (optional)"
          },
          "selectedFights": {
            "type": "number",
            "description": "Number of fights selected for the fight card (optional)"
          }
        },
        "required": [
          "gender",
          "id",
          "lastName",
          "firstName",
          "birthDate",
          "club",
          "license",
          "userId",
          "category",
          "categoryShort"
        ]
      },
      "ModalityErrorType": {
        "type": "string",
        "enum": [
          "WEIGHT",
          "AGE",
          "SAME_CLUB",
          "SAME_ID",
          "OPPOSITE_GENDER",
          "PRIZE_LIST"
        ]
      },
      "ModalityErrorDao": {
        "type": "object",
        "properties": {
          "type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ModalityErrorType"
              }
            ]
          }
        },
        "required": [
          "type"
        ]
      },
      "OpponentGetDto": {
        "type": "object",
        "properties": {
          "gender": {
            "description": "Gender",
            "allOf": [
              {
                "$ref": "#/components/schemas/Gender"
              }
            ]
          },
          "id": {
            "type": "string",
            "description": "Unique boxer ID"
          },
          "lastName": {
            "type": "string",
            "description": "Boxer's last name"
          },
          "firstName": {
            "type": "string",
            "description": "Boxer's first name"
          },
          "birthDate": {
            "type": "string",
            "description": "Date of birth (YYYY-MM-DD)"
          },
          "nbFights": {
            "type": "number",
            "description": "Number of past fights of the boxer"
          },
          "club": {
            "type": "string",
            "description": "Club name"
          },
          "weight": {
            "type": "number",
            "description": "Weight in kg (optional)"
          },
          "license": {
            "type": "string",
            "description": "License number"
          },
          "userId": {
            "type": "string",
            "description": "User who created this boxer"
          },
          "created": {
            "type": "string",
            "description": "Creation timestamp"
          },
          "updated": {
            "type": "string",
            "description": "Update timestamp"
          },
          "category": {
            "type": "string",
            "description": "Weight/age category"
          },
          "categoryShort": {
            "type": "string",
            "description": "Short category label"
          },
          "eligibleFights": {
            "type": "number",
            "description": "Number of eligible fights (optional)"
          },
          "selectedFights": {
            "type": "number",
            "description": "Number of fights selected for the fight card (optional)"
          },
          "weightDifference": {
            "type": "number",
            "description": "Difference in weight between main boxer and opponent"
          },
          "isEligible": {
            "type": "boolean",
            "description": "Whether this opponent is eligible for a fight"
          },
          "fightId": {
            "type": "string",
            "description": "Existing fight ID if a fight already exists"
          },
          "modalityErrors": {
            "description": "Modality errors for this boxer",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModalityErrorDao"
            }
          }
        },
        "required": [
          "gender",
          "id",
          "lastName",
          "firstName",
          "birthDate",
          "club",
          "license",
          "userId",
          "category",
          "categoryShort",
          "weightDifference",
          "isEligible"
        ]
      },
      "FightGetDto": {
        "type": "object",
        "properties": {
          "order": {
            "type": "number",
            "description": "Fight order in the tournament"
          },
          "id": {
            "type": "string",
            "description": "Unique fight ID"
          },
          "boxer1": {
            "description": "Red corner boxer",
            "allOf": [
              {
                "$ref": "#/components/schemas/BoxerGetDto"
              }
            ]
          },
          "boxer2": {
            "description": "Blue corner boxer",
            "allOf": [
              {
                "$ref": "#/components/schemas/BoxerGetDto"
              }
            ]
          },
          "tournamentId": {
            "type": "string",
            "description": "Tournament ID"
          },
          "roundDurationAsSeconds": {
            "type": "number",
            "description": "Fight duration in seconds"
          },
          "rounds": {
            "type": "number",
            "description": "Rounds in the fight"
          }
        },
        "required": [
          "order",
          "id",
          "boxer1",
          "boxer2",
          "tournamentId",
          "roundDurationAsSeconds",
          "rounds"
        ]
      },
...
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}